✨ feat(#14): 위도 및 경도 타입을 Double로 변경, 요청 DTO 필드값 추가

+++ src/main/java/com/backend/domain/store/dto/common/StoreLocationDto.java
@@ -1,9 +1,7 @@
 package com.backend.domain.store.dto.common;
 
-import java.math.BigDecimal;
-
 public record StoreLocationDto(
-	BigDecimal latitude,
-	BigDecimal longitude
+	Double latitude,
+	Double longitude
 ) {
 }
