+++ src/main/java/com/backend/domain/user/dto/common/UserLocationDto.java
@@ -1,9 +1,7 @@
 package com.backend.domain.user.dto.common;
 
-import java.math.BigDecimal;
-
 public record UserLocationDto(
-	BigDecimal latitude,
-	BigDecimal longitude
+	Double latitude,
+	Double longitude
 ) {
 }
